% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queue_add.R
\name{queue_add}
\alias{queue_add}
\title{queue_add}
\usage{
queue_add(
  name,
  date_added = Sys.Date(),
  log_filepath = NULL,
  biometric_lead = NULL,
  user = Sys.getenv("USERNAME"),
  ...
)
}
\arguments{
\item{name}{The name of the project to be added to the queue}

\item{...}{Any additional fields to be filled out}
}
\description{
queue_add
}
\examples{
queue_add('test',
          description = 'A description goes here',
          date_due = as.Date('1992-09-26'),
          something_else = "This doesn't work!")

}
