% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queue_delete.R
\name{queue_delete}
\alias{queue_delete}
\title{queue_delete}
\usage{
queue_delete(name, user = Sys.getenv("USERNAME"), queue_dir = "S:/biometric/")
}
\arguments{
\item{name}{The project to be permanently deleted from the queue}

\item{user}{The name of the user making the deletion}

\item{queue_dir}{The directory containing the queue.rds file}
}
\description{
queue_delete
}
\examples{
queue_delete(name = 'test')
}
