% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queue_load.R
\name{queue_load}
\alias{queue_load}
\title{queue_load}
\usage{
queue_load(queue_dir = "S:/biometric/")
}
\arguments{
\item{queue_dir}{The directory containing the queue.rds file}
}
\value{
The current biometric queue as a data frame
}
\description{
queue_load
}
\examples{
queue <- queue_load()
}
