% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queue_write.R
\name{queue_write}
\alias{queue_write}
\title{Title}
\usage{
queue_write(queue, queue_dir = "S:/biometric/")
}
\arguments{
\item{queue}{The dataframe containing the updated queue}

\item{queue_dir}{The directory containing the queue.rds file}
}
\description{
Title
}
\examples{
queue <- queue_load()
          queue_write(queue)
}
