% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queue_update.R
\name{queue_update}
\alias{queue_update}
\title{queue_update}
\usage{
queue_update(name, user = Sys.getenv("USERNAME"), ...)
}
\arguments{
\item{name}{The name of the project to be updated in the queue}

\item{...}{Any fields to be updated}
}
\description{
queue_update
}
\examples{
queue_update('test',
          description = 'An updated description',
          PI = 'Harry Potter',
          something_else = "This still doesn't work!")

}
